+++ html/menuPage.php
@@ -1,6 +1,6 @@
 <div id="wrapperMenu" class="wrapper">
     <div class="menuPageLinkWrappers col-s-10 col-2" id="menuPageLinkWrapper1" ><a href="analysPage.php">Analys</a></div>
-    <div class="menuPageLinkWrappers col-s-10 col-2" id="menuPageLinkWrapper2"><a href="kassaPage.php">Kassa</a></div>
+    <div class="menuPageLinkWrappers col-s-10 col-2" id="menuPageLinkWrapper2"><a href="html/cashPage.php">Kassa</a></div>
     <div class="menuPageLinkWrappers col-s-10 col-2" id="menuPageLinkWrapper3"><a href="html/depoPage.php">Depo</a></div>
     <div class="menuPageLinkWrappers col-s-10 col-2" id="menuPageLinkWrapper4"><a href="settingsPage.php">Inställningar</a></div>
     <div class="menuPageLinkWrappers col-s-10 col-2" id="menuPageLinkWrapper5"><a href="html/logoutPage.php">Logout</a></div>
